requirements.txt
Flask==2.3.2
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0